requirements.txt
# Typing Extensions (Ensuring Compatibility with TensorFlow & IPython)
typing-extensions==4.5.0  # Downgraded for compatibility with TensorFlow

# Core Libraries
numpy==1.24.3  # Ensuring compatibility with TensorFlow
pandas==1.5.3
matplotlib==3.7.3
seaborn==0.12.2
scikit-learn==1.3.0
imbalanced-learn==0.10.1
datasets==2.14.5
requests==2.31.0

# Machine Learning & Deep Learning
tensorflow==2.13.0
tensorflow-intel==2.13.0  # Ensuring CPU compatibility
torch==2.0.1
transformers==4.33.3
sentence-transformers==2.2.2
keras==2.13.1
huggingface_hub==0.15.1

# NLP and AI Models
nltk==3.8.1
spacy==3.6.1
langdetect==1.0.7
keybert==0.7.0
rake-nltk==1.0.4
yake==0.4.8
thinc==8.1.9  # Ensured compatibility with spaCy and NumPy

# Web Scraping
selenium==4.10.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.2

# Flask & FastAPI Backend
flask==3.0.0
flask-cors==4.0.0
fastapi==0.100.0
uvicorn==0.22.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Database (MongoDB)
pymongo==4.6.0
pydantic==1.10.9

# Jupyter Notebook & Kernel (Fixed IPython Conflict)
notebook==6.5.6  # Changed from 7.0.0 to avoid JupyterLab conflicts
ipykernel==6.25.2
ipython==8.12.3  # Ensured compatibility

# Visualization & Utility
wordcloud==1.9.2
shap==0.43.0  # Added since it required `numba`
numba==0.57.0  # Downgraded for better compatibility with NumPy 1.24.3
lime==0.2.0.1

# Stopwords & NLP Extras
spacy-langdetect==0.1.2
spacy-lookups-data==1.0.5
